test: User 도메인 회원가입 로직 리팩토링 후 실패하는 테스트 수정
[SW-30]
